feat: Add KSP2PrometheusExporter project with basic functionality
This commit adds the KSP2PrometheusExporter project, which exports telemetry data to Prometheus. It uses the SpaceWarp API and the Prometheus.NET library. Currently, it exports the orbital speed of the active vessel vehicle as a gauge to port 9102. This functionality is not yet integrated into the UI or controlled by user input.

Disclaimer: This message has been generated with gpt-commit-messages: https://github.com/Goddchen/gpt-commit-messages
Signed-off-by: Goddchen <goddchen@gmail.com>